03-04-2015
* много всяких изменений
* у Тараса нет защиты от неполно принятого блока - необходимо передавать в разборщик ещё и реально считанную длину, чтобы не выходить за пределы памяти

+++ publicclass.cpp
@@ -150,7 +150,7 @@ publicclass::DataRec *publicclass::FindElem(DataRec *dr, qint16 id)
   for(;;dr++)
   {
     if(dr->id==id) return dr;
-    if(dr->id==0xFFFF) break;
+    if(dr->id==static_cast<qint16>(0xFFFF)) break;
   }
   return NULL;
 }
